Updates to stereo cameras.
Stereo cameras are now attached to eyes on the geometry model
and oriented properly.

Fog settings in the scene were also prevening distant objects from
appearing in the stereo cameras. This could possibly be due to some
scaling mismatch between the global camera and the imported geometry
model to which the cameras are attached. Increasing the fog parameters
now allows more distant objects to be seen. I may consider removing
the fog effect altogether.